[Feat] 공지사항 조회수 집계 로직 수정 (#659)
* feat: Notice 도메인에 조회수 필드 추가

* chore: flyway 파일 추가

* feat: 트랜잭션 고려해서 조회수 증가 로직 분리

* fix: 조회 수 증가를 완전 새로운 트랜잭션으로 수행하도록 변경

* chore: 사용하지 않는 도메인 메서드 삭제

+++ src/main/java/com/umc/product/notice/adapter/in/web/dto/response/query/GetNoticeSummaryResponse.java
@@ -17,7 +17,7 @@ public record GetNoticeSummaryResponse(
     String content,
     // UI에서 빨간 점 표시 용; 알람을 전송하는 것인지에 대한 내용
     Boolean shouldSendNotification,
-    Integer viewCount,
+    Long viewCount,
     Instant createdAt,
     NoticeTargetInfo targetInfo,
     Long authorChallengerId,
