+++ src/main/java/com/umc/product/notice/application/port/in/query/dto/NoticeSummary.java
@@ -8,7 +8,7 @@ public record NoticeSummary(
     String title,
     String content,
     Boolean shouldSendNotification,
-    Integer viewCount,
+    Long viewCount,
     Instant createdAt,
     NoticeTargetInfo targetInfo,
     Long authorMemberId,
